feat(ui-state): implement batch retrieval of UI state records
- Added `get_ui_state_batch` method to `UiStateStore` trait for fetching multiple UI state records by keys.
- Implemented batch retrieval logic in `AnyStore` with SQL query optimization.
- Introduced `UiStateBatchEntryResponse` and `UiStateBatchResponse` models for structured API responses.
- Updated API definitions to include new batch endpoint `/v1/ui-state` with query parameters.
- Enhanced client-side storage to coalesce concurrent `getItem` calls into a single batched request.
- Added tests to verify batch retrieval functionality and ensure correct handling of present and missing keys.
- Updated Python SDK to support new batch API, including request and response models.

+++ crates/slab-app-core/src/infra/db/repository/media_task.rs
@@ -204,7 +204,7 @@ impl MediaTaskStore for AnyStore {
         .bind(i64::from(video_task.width))
         .bind(i64::from(video_task.height))
         .bind(i64::from(video_task.frames))
-        .bind(f64::from(video_task.fps))
+        .bind(video_task.fps)
         .bind(&video_task.reference_image_path)
         .bind(&video_task.request_data)
         .bind(video_task.created_at.to_rfc3339())
